requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
pandas==2.2.2
numpy==1.26.4
sqlglot==25.17.0
python-dotenv==1.0.1
openai>=1.44.0